src/lib/external/requirements.txt
# Core dependencies
numpy>=1.20.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=8.0.0
tensorflow-macos>=2.15.0
tensorflow-metal>=1.0.0

# PyKTok dependencies
selenium>=4.0.0
beautifulsoup4>=4.9.0
pandas>=1.0.0
requests>=2.25.0
chromedriver-autoinstaller>=0.2.2
python-dotenv>=0.19.0

# FAISS dependencies
faiss-cpu>=1.7.0
scikit-learn>=0.24.0

# Video-Analyzer dependencies
opencv-python>=4.5.0
moviepy>=1.0.0
pytube>=10.0.0
ffmpeg-python>=0.2.0

# CLIP dependencies
ftfy>=5.8
regex>=2020.0
tqdm>=4.0.0
transformers>=4.0.0

# OSINT Tools dependencies
tweepy>=4.0.0
instaloader>=4.9
python-dotenv>=0.19.0
google-api-python-client>=2.0.0
youtube-dl>=2021.12.17

# Perplexica UI dependencies
streamlit>=1.0.0
plotly>=5.0.0
altair>=4.0.0
st-annotated-text>=2.0.0
streamlit-extras>=0.2.0

# FastAPI dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
aiofiles>=0.7.0
httpx>=0.23.0

# Development dependencies
pytest>=6.0.0
black>=21.0.0
isort>=5.0.0
mypy>=0.900 